assignment-3/pyproject.toml
[tool.poetry]
name = "ds561-a3"
version = "0.1.0"
description = ""
authors = ["Alicja Mahr <bmahr@bu.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
certifi = "^2023.7.22"
functions-framework = "^3.4.0"
google-cloud-storage = "^2.11.0"
google-cloud-pubsub = "^2.18.4"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
